pyproject.toml
[project]
name = "class-8-assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.10",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "tavily-python>=0.7.11",
]
